fix(warnings): resolve MSVC C4244/C4100/C4189 warnings
- sendinput_input: cast MapVirtualKeyA result to WORD
- win32_window: mark unused monitorIndex parameter
- win32_screen: remove unused index variables
- win32_process: mark unused force parameter
- win32_system: suppress unused type variable
- win32_recorder: cast uint64_t timestamps to DWORD

+++ lib/wingman/src/platform/win/sendinput_input.cpp
@@ -166,7 +166,7 @@ public:
                 continue;
             }
 
-            WORD scanCode = MapVirtualKeyA(LOBYTE(vk), 0);
+            WORD scanCode = static_cast<WORD>(MapVirtualKeyA(LOBYTE(vk), 0));
 
             INPUT inputs[2] = {};
             inputs[0].type = INPUT_KEYBOARD;
